Update build.nim, try to fulfill fork's purpose
This is meant to fix an issue with the latest iteration of the Reimers Lab MABE workflow. Using mbuild to compile our fork in debug mode caused a compile-time error with H5open undefined. We're using HDF5 with armadillo to save connectome data at the end of the simulations, to be later processed in a separate program and have relevant information subsequently brought into MATLAB for further analysis and visualization. The debug mode is because there is currently an unpredictable segfault issue in our MABE simulations, and the simple gdb output without added debuginfo was not adequate for solving the problem. This may or may not fix the mbuild issue; has not been tested yet.